Organized syntax simulator into modules
