+++ quasim/src/syntax_simulator.rs
@@ -1,790 +1,16 @@
-use core::f32;
-use std::{
-    collections::VecDeque,
-    fmt::{Debug, Display},
-    ops::{Add, Mul, Neg},
-    vec,
-};
+mod basis;
+mod scalar;
+mod state;
+#[cfg(test)]
+mod tests;
 
-use log::trace;
-use nalgebra::{Complex, DVector};
 use rand::random;
 
 use crate::{
     circuit::{Circuit, PureCircuit, pc::CircuitPc},
-    gate::{Gate, GateType, QBits},
     instruction::PureInstruction,
+    syntax_simulator::state::{ScaledState, SumOfScaledStates},
 };
-
-#[derive(Clone, PartialEq, Copy)]
-struct Scalar {
-    frac_1_sqrt_2_power: usize,
-    number: Complex<f32>,
-}
-
-impl Scalar {
-    const ZERO: Scalar = Scalar {
-        frac_1_sqrt_2_power: 0,
-        number: Complex::new(0.0, 0.0),
-    };
-
-    const ONE: Scalar = Scalar {
-        frac_1_sqrt_2_power: 0,
-        number: Complex::new(1.0, 0.0),
-    };
-
-    const I: Scalar = Scalar {
-        frac_1_sqrt_2_power: 0,
-        number: Complex::new(0.0, 1.0),
-    };
-
-    const FRAC_1_SQRT_2: Scalar = Scalar {
-        frac_1_sqrt_2_power: 1,
-        number: Complex::new(1.0, 0.0),
-    };
-
-    fn mul_frac_1_sqrt_2(&mut self) {
-        self.frac_1_sqrt_2_power += 1;
-    }
-
-    fn is_zero(&self) -> bool {
-        self == &Scalar::ZERO
-    }
-
-    fn probability(&self) -> f32 {
-        let c: Complex<f32> = (*self).into();
-        c.norm_sqr()
-    }
-}
-
-impl Add for Scalar {
-    type Output = Self;
-
-    fn add(self, rhs: Self) -> Self::Output {
-        if self.frac_1_sqrt_2_power == rhs.frac_1_sqrt_2_power {
-            Self {
-                frac_1_sqrt_2_power: self.frac_1_sqrt_2_power,
-                number: self.number + rhs.number,
-            }
-        } else {
-            Self {
-                frac_1_sqrt_2_power: self.frac_1_sqrt_2_power,
-                number: self.number + Complex::<f32>::from(rhs),
-            }
-        }
-    }
-}
-
-impl Neg for Scalar {
-    type Output = Self;
-
-    fn neg(self) -> Self::Output {
-        Self {
-            frac_1_sqrt_2_power: self.frac_1_sqrt_2_power,
-            number: -self.number,
-        }
-    }
-}
-
-impl Mul for Scalar {
-    type Output = Self;
-
-    fn mul(self, rhs: Self) -> Self::Output {
-        Self {
-            frac_1_sqrt_2_power: self.frac_1_sqrt_2_power + rhs.frac_1_sqrt_2_power,
-            number: self.number * rhs.number,
-        }
-    }
-}
-
-impl Mul<ExtendedBasis> for Scalar {
-    type Output = ScaledState;
-
-    fn mul(self, rhs: ExtendedBasis) -> Self::Output {
-        ScaledState(rhs, self)
-    }
-}
-
-impl Mul<ScaledState> for Scalar {
-    type Output = ScaledState;
-
-    fn mul(self, rhs: ScaledState) -> Self::Output {
-        let ScaledState(state, scalar) = rhs;
-        ScaledState(state, self * scalar)
-    }
-}
-
-impl Mul<Sum> for Scalar {
-    type Output = Sum;
-
-    fn mul(self, rhs: Sum) -> Sum {
-        rhs.into_iter().map(|ss| self * ss).collect()
-    }
-}
-
-impl Debug for Scalar {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self)
-    }
-}
-
-impl Display for Scalar {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        if self.frac_1_sqrt_2_power == 0 {
-            write!(f, "{}", self.number)
-        } else {
-            write!(f, "({} / √2^{}))", self.number, self.frac_1_sqrt_2_power)
-        }
-    }
-}
-
-impl From<Complex<f32>> for Scalar {
-    fn from(value: Complex<f32>) -> Self {
-        Scalar {
-            frac_1_sqrt_2_power: 0,
-            number: value,
-        }
-    }
-}
-
-impl From<Scalar> for Complex<f32> {
-    fn from(value: Scalar) -> Self {
-        value.number * f32::consts::FRAC_1_SQRT_2.powi(value.frac_1_sqrt_2_power as i32)
-    }
-}
-
-type TrueQubitBasis = bool;
-
-#[derive(Copy, Clone)]
-struct ScaledQubitBasis(TrueQubitBasis, Scalar);
-
-type ScaledQubit = [ScaledQubitBasis; 2];
-
-impl Debug for ScaledQubitBasis {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}|{}⟩", self.1, if self.0 { 1 } else { 0 })
-    }
-}
-
-#[derive(Clone, Copy, PartialEq)]
-enum ExtendedQubitBasis {
-    Zero,
-    One,
-    Plus,
-    Minus,
-    I,
-    MinusI,
-}
-
-impl Debug for ExtendedQubitBasis {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        use ExtendedQubitBasis::*;
-        let s = match self {
-            Zero => "0",
-            One => "1",
-            Plus => "+",
-            Minus => "-",
-            I => "i",
-            MinusI => "(-i)",
-        };
-        write!(f, "{}", s)
-    }
-}
-
-impl ExtendedQubitBasis {
-    fn semantic_expressions(&self) -> ScaledQubit {
-        use ExtendedQubitBasis::*;
-        match self {
-            Zero => [
-                ScaledQubitBasis(false, Scalar::ONE),
-                ScaledQubitBasis(false, Scalar::ZERO),
-            ],
-            One => [
-                ScaledQubitBasis(true, Scalar::ONE),
-                ScaledQubitBasis(true, Scalar::ZERO),
-            ],
-            Plus => [
-                ScaledQubitBasis(false, Scalar::FRAC_1_SQRT_2),
-                ScaledQubitBasis(true, Scalar::FRAC_1_SQRT_2),
-            ],
-            Minus => [
-                ScaledQubitBasis(false, Scalar::FRAC_1_SQRT_2),
-                ScaledQubitBasis(true, -Scalar::FRAC_1_SQRT_2),
-            ],
-            I => [
-                ScaledQubitBasis(false, Scalar::FRAC_1_SQRT_2),
-                ScaledQubitBasis(true, Scalar::FRAC_1_SQRT_2 * Scalar::I),
-            ],
-            MinusI => [
-                ScaledQubitBasis(false, Scalar::FRAC_1_SQRT_2),
-                ScaledQubitBasis(true, -Scalar::FRAC_1_SQRT_2 * Scalar::I),
-            ],
-        }
-    }
-
-    // Gates
-
-    #[inline(always)]
-    fn x(&self) -> ExtendedQubitBasis {
-        use ExtendedQubitBasis::*;
-        match self {
-            Zero => One,
-            One => Zero,
-            Plus => Plus,
-            Minus => Minus,
-            I => MinusI,
-            MinusI => I,
-        }
-    }
-
-    #[inline(always)]
-    fn y(&self) -> ExtendedQubitBasis {
-        use ExtendedQubitBasis::*;
-        match self {
-            Zero => One,
-            One => Zero,
-            Plus => Minus,
-            Minus => Plus,
-            I => I,
-            MinusI => MinusI,
-        }
-    }
-
-    fn z(&self) -> ExtendedQubitBasis {
-        use ExtendedQubitBasis::*;
-        match self {
-            Zero => Zero,
-            One => One,
-            Plus => Minus,
-            Minus => Plus,
-            I => MinusI,
-            MinusI => I,
-        }
-    }
-
-    #[inline(always)]
-    fn h(&self) -> ExtendedQubitBasis {
-        use ExtendedQubitBasis::*;
-        match self {
-            Zero => Plus,
-            One => Minus,
-            Plus => Zero,
-            Minus => One,
-            I => MinusI,
-            MinusI => I,
-        }
-    }
-
-    #[inline(always)]
-    fn s(&self) -> ExtendedQubitBasis {
-        // The S gate is just a 90 degree rotation around z-axis on bloch sphere.
-        use ExtendedQubitBasis::*;
-        match self {
-            Zero => Zero,
-            One => One,
-            Plus => I,
-            I => Minus,
-            Minus => MinusI,
-            MinusI => Plus,
-        }
-    }
-}
-
-#[derive(Clone)]
-enum ExtendedBasis {
-    Binary(usize),
-    Superposition(Vec<ExtendedQubitBasis>),
-}
-
-impl ExtendedBasis {
-    /// Expands the given qubit in the basis, returning the two cases of the expansion if
-    /// the qubit is in an extended basis (eg, |+⟩, |−⟩, |i⟩, |−i⟩),
-    /// otherwise returning the same basis if the qubit is already binary (eg, |0⟩, |1⟩).
-    fn expand_qubit(self, qubit_to_expand: usize) -> [Option<ScaledState>; 2] {
-        use ExtendedBasis::*;
-        match self {
-            Binary(_) => [Some(ScaledState(self, Scalar::ONE)), None],
-            Superposition(bases) => {
-                let qubit_basis_to_expand = bases
-                    .get(qubit_to_expand)
-                    .expect("Tried to expand a qubit out of bounds");
-
-                use ExtendedQubitBasis::*;
-                match *qubit_basis_to_expand {
-                    One | Zero => {
-                        // If the qubit is already in a binary state, there is nothing to expand.
-                        return [
-                            Some(ScaledState(Self::Superposition(bases), Scalar::ONE)),
-                            None,
-                        ];
-                    }
-                    _ => (),
-                };
-
-                let apply_minus: bool = matches!(
-                    qubit_basis_to_expand,
-                    ExtendedQubitBasis::Minus | ExtendedQubitBasis::MinusI
-                );
-
-                let mut zero_case = bases.clone();
-                zero_case[qubit_to_expand] = Zero;
-                let mut one_case = bases;
-                one_case[qubit_to_expand] = One;
-
-                [
-                    Some(ScaledState(
-                        Self::Superposition(zero_case),
-                        Scalar::FRAC_1_SQRT_2,
-                    )),
-                    Some(ScaledState(
-                        Self::Superposition(one_case),
-                        if apply_minus {
-                            -Scalar::FRAC_1_SQRT_2
-                        } else {
-                            Scalar::FRAC_1_SQRT_2
-                        },
-                    )),
-                ]
-            }
-        }
-    }
-
-    /// Expands all the qubit indexes in the given list, meaning that qubits in an extended
-    /// basis (eg, |+⟩, |−⟩, |i⟩, |−i⟩) will be expanded into a sum of binary states,
-    /// while qubits already in a binary basis (eg, |0⟩, |1⟩) will be left unchanged.
-    fn expand_qubits(self, qubits_to_expand: QBits) -> Sum {
-        self.expand_qubits_helper(qubits_to_expand.get_indices().into())
-    }
-
-    fn expand_qubits_helper(self, mut qubits_to_expand: VecDeque<usize>) -> Sum {
-        let Some(qubit_to_expand) = qubits_to_expand.pop_front() else {
-            return vec![ScaledState(self, Scalar::ONE)];
-        };
-
-        use ExtendedBasis::*;
-        match self {
-            Binary(_) => vec![ScaledState(self, Scalar::ONE)],
-            Superposition(_) => {
-                // Expand the first qubit, which yields just two terms.
-                let possible_expansion = self.clone().expand_qubit(qubit_to_expand);
-
-                // Actually, expanding |0⟩ or |1⟩ just gives one term, so we need to filter out the None case.
-                let just_existing_expansions = possible_expansion
-                    .iter()
-                    // Flatten removes the Option wrapping
-                    .flatten();
-
-                // At this point we have all the expansions for the first qubit to expand.
-                // Take those terms we expanded into, and just run this function on those terms,
-                // but now with one less qubit to expand.
-
-                let all_terms = just_existing_expansions
-                    .flat_map(|ScaledState(basis, scalar)| {
-                        // If a scaled state turns into more terms, the scalar needs to be distributed across the terms.
-                        *scalar * basis.clone().expand_qubits_helper(qubits_to_expand.clone()) // Expensive clone(s)?
-                    })
-                    .collect::<Sum>();
-
-                all_terms
-            }
-        }
-    }
-
-    fn try_into_binary(&self) -> Option<usize> {
-        match self {
-            ExtendedBasis::Binary(bitstring) => Some(*bitstring),
-            ExtendedBasis::Superposition(_) => None,
-        }
-    }
-
-    fn into_binary(&self) -> usize {
-        self.try_into_binary()
-            .expect("Cannot convert superposition basis into binary basis!")
-    }
-
-    fn vec_form(self) -> Vec<ExtendedQubitBasis> {
-        match self {
-            ExtendedBasis::Binary(mut bitstring) => {
-                let mut bases = Vec::new();
-                while bitstring > 0 {
-                    bases.push(ExtendedQubitBasis::One);
-                    bitstring -= 1;
-                }
-                bases
-            }
-            ExtendedBasis::Superposition(bases) => bases,
-        }
-    }
-
-    fn vec_form_padded_to_len(self, n: usize) -> Vec<ExtendedQubitBasis> {
-        let mut vec_form = self.vec_form();
-        vec_form.resize(n, ExtendedQubitBasis::Zero);
-        vec_form
-    }
-
-    fn pad_to_length(superposition: &mut Vec<ExtendedQubitBasis>, n: usize) {
-        if superposition.len() < n {
-            superposition.resize(n, ExtendedQubitBasis::Zero);
-        }
-    }
-
-    fn all_inherent_states_of(extended_basis: ExtendedBasis) -> Sum {
-        let mut extended_basis = match extended_basis {
-            ExtendedBasis::Binary(bitstring) => {
-                return vec![ScaledState(ExtendedBasis::Binary(bitstring), Scalar::ONE)];
-            }
-            ExtendedBasis::Superposition(b) => b,
-        };
-
-        let Some(qubit_basis) = extended_basis.pop() else {
-            panic!("Cannot get inherent states of empty basis!");
-        };
-
-        if extended_basis.is_empty() {
-            return qubit_basis
-                .semantic_expressions()
-                .iter()
-                .filter(|ScaledQubitBasis(_, scalar)| !scalar.is_zero())
-                .map(|ScaledQubitBasis(basis, scalar)| {
-                    let basis = if *basis { 1 } else { 0 };
-                    ScaledState(ExtendedBasis::Binary(basis), *scalar)
-                })
-                .collect();
-        }
-
-        let semantics = qubit_basis.semantic_expressions();
-
-        semantics
-            .into_iter()
-            .filter(|ScaledQubitBasis(_, s)| !s.is_zero())
-            .map(|msb| {
-                let lesser_bits = Self::all_inherent_states_of(ExtendedBasis::Superposition(
-                    extended_basis.clone(),
-                ));
-                lesser_bits
-                    .iter()
-                    .map(|lsb| {
-                        let lsb_state = match &lsb.0 {
-                            ExtendedBasis::Binary(bitstring) => bitstring,
-                            ExtendedBasis::Superposition(_) => {
-                                panic!("Lesser bits should always be binary at this point!")
-                            }
-                        };
-                        let basis = (msb.0 as usize) << extended_basis.len() | lsb_state;
-                        let scalar_product = msb.1 * lsb.1;
-                        trace!("Basis: {}, Scalar: {}", basis, scalar_product);
-                        ScaledState(ExtendedBasis::Binary(basis), scalar_product)
-                    })
-                    .collect::<Vec<_>>()
-            })
-            .flatten()
-            .collect()
-    }
-
-    pub fn all_inherent_states(self: ExtendedBasis) -> Sum {
-        Self::all_inherent_states_of(self.clone())
-    }
-
-    // Gates
-
-    fn x(&mut self, target: usize) {
-        use ExtendedBasis::*;
-        match self {
-            Binary(bits) => *bits = *bits ^ (1 << target),
-            Superposition(bases) => {
-                Self::pad_to_length(bases, target + 1);
-                bases[target] = bases[target].x();
-            }
-        }
-    }
-
-    fn y(&mut self, target: usize) {
-        use ExtendedBasis::*;
-        match self {
-            Binary(bits) => *bits = *bits ^ (1 << target),
-            Superposition(bases) => {
-                Self::pad_to_length(bases, target + 1);
-                bases[target] = bases[target].y();
-            }
-        }
-    }
-
-    fn z(&mut self, target: usize) {
-        use ExtendedBasis::*;
-        match self {
-            Binary(_) => (),
-            Superposition(bases) => {
-                Self::pad_to_length(bases, target + 1);
-                bases[target] = bases[target].z();
-            }
-        }
-    }
-
-    fn h(&mut self, target: usize) {
-        use ExtendedBasis::*;
-        match self {
-            Binary(bits) => {
-                let bit_is_one = (*bits & (1 << target)) != 0;
-                let new_qubit_basis = if bit_is_one {
-                    ExtendedQubitBasis::Minus
-                } else {
-                    ExtendedQubitBasis::Plus
-                };
-                // Clone is cheap because self is effectively a usize
-                let mut new_bases = self.clone().vec_form_padded_to_len(target + 1);
-                new_bases[target] = new_qubit_basis;
-                *self = Superposition(new_bases);
-            }
-            Superposition(bases) => {
-                Self::pad_to_length(bases, target + 1);
-                bases[target] = bases[target].h();
-            }
-        }
-    }
-
-    fn s(&mut self, target: usize) {
-        use ExtendedBasis::*;
-        match self {
-            Binary(_) => (),
-            Superposition(bases) => {
-                Self::pad_to_length(bases, target + 1);
-                bases[target] = bases[target].s();
-            }
-        }
-    }
-
-    fn r_y(self, target: usize, angle: f32) -> [ScaledState; 2] {
-        use ExtendedBasis::*;
-        match self {
-            Binary(_) => Self::r_y(Superposition(self.vec_form()), target, angle),
-            Superposition(_) => {
-                let angle = Complex::from(angle / 2f32);
-                let expanded = self.expand_qubit(target);
-
-                let Some(ScaledState(ref zero_basis, ref zero_scalar)) = expanded[0] else {
-                    panic!("Would expect at least one state after expansion");
-                };
-
-                let top_left = Scalar::from(angle.cos()) * *zero_scalar;
-                let bottom_left = Scalar::from(angle.sin()) * *zero_scalar;
-
-                let r: [ScaledState; 2] =
-                    if let Some(ScaledState(ref one_basis, ref one_scalar)) = expanded[0] {
-                        let top_right = Scalar::from(-angle.sin()) * *one_scalar;
-                        let bottom_right = Scalar::from(angle.cos()) * *one_scalar;
-
-                        [
-                            (top_left + top_right) * zero_basis.clone(),
-                            (bottom_left + bottom_right) * one_basis.clone(),
-                        ]
-                    } else {
-                        [
-                            top_left * zero_basis.clone(),
-                            bottom_left * zero_basis.clone(),
-                        ]
-                    };
-
-                r
-            }
-        }
-    }
-
-    fn r_z(self, target: usize, angle: f32) -> [Option<ScaledState>; 2] {
-        use ExtendedBasis::*;
-        match self {
-            Binary(_) => [Some(ScaledState(self, Scalar::ONE)), None],
-            Superposition(_) => {
-                let mut result = self.expand_qubit(target);
-
-                let Some(Some(ScaledState(_, one_case_scalar))) = result.get_mut(1) else {
-                    // If the qubit was already in a binary state, the phase cannot be altered.
-                    return result;
-                };
-
-                let phase_change: Scalar = Complex::exp(Complex::I * angle).into();
-                *one_case_scalar = phase_change * *one_case_scalar;
-
-                result
-            }
-        }
-    }
-}
-
-impl From<ScaledQubitBasis> for ScaledState {
-    fn from(value: ScaledQubitBasis) -> Self {
-        let basis = match value.0 {
-            false => 0,
-            true => 1,
-        };
-        ScaledState(ExtendedBasis::Binary(basis), value.1)
-    }
-}
-
-impl Debug for ExtendedBasis {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ExtendedBasis::Binary(bitstring) => write!(f, "|{}⟩", bitstring),
-            ExtendedBasis::Superposition(bases) => {
-                write!(f, "|")?;
-                for basis in bases {
-                    write!(f, "{:?}", basis)?;
-                }
-                write!(f, "⟩")?;
-                Ok(())
-            }
-        }
-    }
-}
-
-impl PartialEq for ExtendedBasis {
-    fn eq(&self, other: &Self) -> bool {
-        use ExtendedBasis::*;
-        match (self, other) {
-            (Binary(b1), Binary(b2)) => b1 == b2,
-            (Superposition(s1), Superposition(s2)) => s1 == s2,
-            (Binary(l), Superposition(r)) => {
-                let mut l = *l;
-                for basis in r {
-                    let bit_is_one = (l & 1) == 1;
-                    let expected_basis = if bit_is_one {
-                        ExtendedQubitBasis::One
-                    } else {
-                        ExtendedQubitBasis::Zero
-                    };
-                    if *basis != expected_basis {
-                        return false;
-                    }
-                    l = l >> 1;
-                }
-                true
-            }
-            _ => other == self,
-        }
-    }
-}
-
-#[derive(Clone, PartialEq)]
-struct ScaledState(ExtendedBasis, Scalar);
-
-impl ScaledState {
-    fn all_inherent_states(&self) -> Sum {
-        let my_scalar = self.1;
-
-        self.0
-            .clone()
-            .all_inherent_states()
-            .iter()
-            // Guaranteed usize means cheap clone.
-            .map(|substate| my_scalar * substate.clone())
-            .collect()
-    }
-}
-
-impl Debug for ScaledState {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let ScaledState(basis, scalar) = self;
-        write!(f, "{}{:?}", scalar, basis)
-    }
-}
-
-impl From<(usize, Scalar)> for ScaledState {
-    fn from(value: (usize, Scalar)) -> Self {
-        let basis = value.0;
-        let scalar = value.1;
-        ScaledState(ExtendedBasis::Binary(basis), scalar)
-    }
-}
-
-type Sum = Vec<ScaledState>;
-
-#[derive(Debug, Clone)]
-struct SumOfScaledStates {
-    n_qubits: usize,
-    sum: Sum,
-}
-
-impl SumOfScaledStates {
-    fn guaranteed_full_zero(n_qubits: usize) -> Self {
-        let n_zeros = ScaledState(ExtendedBasis::Binary(0), Scalar::ONE);
-        Self {
-            n_qubits,
-            sum: vec![n_zeros],
-        }
-    }
-
-    /// Returns the probability distribution over all basis states
-    /// that exist at this point. Extended bases (eg, |+⟩, |−⟩, |i⟩, |−i⟩)
-    /// will not be expanded, so the resulting states may still contain extended bases.
-    /// If you want to get the distribution over only binary states,
-    /// use `unsugared_probability_distribution` instead.
-    fn probability_distribution(&self) -> impl Iterator<Item = &ScaledState> {
-        self.sum.iter()
-    }
-
-    pub fn apply_controlled_gate(
-        &self,
-        unconditional_single_qubit_gate: fn(&mut ExtendedBasis, usize),
-        controls: QBits,
-        target: usize,
-    ) -> Sum {
-        let mut result = self.expand_necessary_controls(controls);
-
-        result
-            .iter_mut()
-            // Keep only the states where all the controls are satisfied
-            .filter(|ScaledState(basis, _)| {
-                use ExtendedBasis::*;
-                match basis {
-                    Binary(bits) => bits & controls.get_bitstring() == controls.get_bitstring(),
-                    Superposition(bases) => {
-                        use ExtendedQubitBasis::*;
-                        controls
-                            .get_indices()
-                            .iter()
-                            .all(|&i| bases.get(i) == Some(&One))
-                    }
-                }
-            })
-            .for_each(|ScaledState(basis, _)| {
-                unconditional_single_qubit_gate(basis, target);
-            });
-
-        result
-    }
-
-    /// Expands only states that are necessary to check the controls,
-    /// meaning that states where any control qubit is definite zero will be ignored,
-    /// while others will be expanded so checks can be performed on the underlying binary states.
-    fn expand_necessary_controls(&self, controls: QBits) -> Sum {
-        let control_indices = controls.get_indices();
-        let expanded: Sum = self
-            .sum
-            .iter()
-            .map(|ScaledState(basis, scalar)| {
-                use ExtendedBasis::*;
-                match basis {
-                    Superposition(bases) => {
-                        let any_controls_are_zero = control_indices.iter().any(|&i| {
-                            use ExtendedQubitBasis::*;
-                            bases.get(i).unwrap_or(&Zero) == &Zero
-                        });
-
-                        if any_controls_are_zero {
-                            return vec![ScaledState(Superposition(bases.clone()), *scalar)];
-                        } else {
-                            // A term where the controls are fulfilled is contained within `basis`
-                            basis.clone().expand_qubits(controls) // Probably expensive clone
-                        }
-                    }
-                    Binary(bits) => vec![ScaledState(Binary(*bits), *scalar)],
-                }
-            })
-            .flatten()
-            .collect();
-
-        expanded
-    }
-}
 
 pub struct SyntaxSimulator {
     circuit: Circuit<PureCircuit>,
@@ -799,100 +25,20 @@ pub enum SyntaxSimError {
 }
 
 impl SyntaxSimulator {
-    fn apply_gate(expr: &mut SumOfScaledStates, gate: &Gate) {
-        let controls = gate.get_control_bits();
-        let targets = gate.get_target_bits();
-        assert_eq!(targets.get_indices().len(), 1);
-        let target = targets.get_indices()[0];
-
-        trace!(
-            "Applying gate {:?} with controls {:?} and target {}",
-            gate.get_type(),
-            controls,
-            target
-        );
-
-        use GateType::*;
-        expr.sum = match gate.get_type() {
-            X => expr.apply_controlled_gate(ExtendedBasis::x, controls, target),
-            Y => expr.apply_controlled_gate(ExtendedBasis::y, controls, target),
-            Z => expr.apply_controlled_gate(ExtendedBasis::z, controls, target),
-            H => expr.apply_controlled_gate(ExtendedBasis::h, controls, target),
-            S => expr.apply_controlled_gate(ExtendedBasis::s, controls, target),
-            SWAP => {
-                panic!("SWAP gates should have been removed in the circuit preprocessing step!")
-            }
-            U(theta, phi, lambda) => {
-                let is_just_phase_change = theta == 0f64;
-                let mut states_to_alter: Vec<ScaledState> = Vec::new();
-                let mut sum = expr.expand_necessary_controls(controls);
-                sum.retain(|ss| {
-                    let ScaledState(basis, _) = ss;
-                    use ExtendedBasis::*;
-                    match basis {
-                        Binary(bits) => bits & controls.get_bitstring() == controls.get_bitstring(),
-                        Superposition(bases) => {
-                            if is_just_phase_change && matches!(bases[target], Zero | One) {
-                                // The zero and one states are unaffected by phase changes. Just retain the state.
-                                return true;
-                            }
-
-                            use ExtendedQubitBasis::*;
-                            let all_controls_set = controls
-                                .get_indices()
-                                .iter()
-                                .all(|&i| bases.get(i) == Some(&One));
-
-                            if all_controls_set {
-                                states_to_alter.push(ss.clone());
-                                return false;
-                            } else {
-                                return true;
-                            }
-                        }
-                    }
-                });
-
-                let new_states = states_to_alter
-                    .iter()
-                    .flat_map(|ScaledState(basis, _)| basis.clone().r_z(target, lambda as f32))
-                    .flatten()
-                    .map(|ss| ss.clone())
-                    .collect::<Vec<_>>();
-
-                let new_states = new_states
-                    .iter()
-                    .flat_map(|ScaledState(basis, _)| basis.clone().r_y(target, theta as f32))
-                    .map(|ss| ss.clone())
-                    .collect::<Vec<_>>();
-
-                let new_states = new_states
-                    .iter()
-                    .flat_map(|ScaledState(basis, _)| basis.clone().r_z(target, phi as f32))
-                    .flatten()
-                    .map(|ss| ss.clone())
-                    .collect::<Vec<_>>();
-
-                sum.extend(new_states);
-                sum
-            }
-        };
-    }
-
     fn step(&mut self) -> Option<&SumOfScaledStates> {
         let Some(instruction) = self.circuit.instruction(&self.pc) else {
             return None;
         };
         match instruction {
             PureInstruction::Gate(gate) => {
-                Self::apply_gate(&mut self.state, &gate);
+                self.state.apply_gate(&gate);
                 self.pc.increment();
             }
             PureInstruction::Call(_, _, _) => {
                 todo!();
             }
         }
-        return Some(&self.state);
+        Some(&self.state)
     }
 
     fn step_all(&mut self) -> &SumOfScaledStates {
@@ -945,75 +91,5 @@ impl TryFrom<Circuit<PureCircuit>> for SyntaxSimulator {
             pc: CircuitPc::default(),
             state: SumOfScaledStates::guaranteed_full_zero(n_qubits),
         })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::simulator::BuildSimulator;
-
-    use super::*;
-
-    #[test]
-    fn test_expand_qubits() {
-        let basis_to_expand =
-            ExtendedBasis::Superposition(vec![ExtendedQubitBasis::Plus, ExtendedQubitBasis::Minus]);
-
-        use ExtendedQubitBasis::*;
-        assert_eq!(
-            basis_to_expand.clone().expand_qubit(0),
-            [
-                Some(ScaledState(
-                    ExtendedBasis::Superposition(vec![Zero, Minus]),
-                    Scalar::FRAC_1_SQRT_2
-                )),
-                Some(ScaledState(
-                    ExtendedBasis::Superposition(vec![One, Minus]),
-                    Scalar::FRAC_1_SQRT_2
-                )),
-            ]
-        );
-
-        assert_eq!(
-            basis_to_expand.clone().expand_qubit(1),
-            [
-                Some(ScaledState(
-                    ExtendedBasis::Superposition(vec![Plus, Zero]),
-                    Scalar::FRAC_1_SQRT_2
-                )),
-                Some(ScaledState(
-                    ExtendedBasis::Superposition(vec![Plus, One]),
-                    -Scalar::FRAC_1_SQRT_2
-                )),
-            ]
-        );
-
-        let expand_both = QBits::from_indices(&[0, 1]);
-        let half = Scalar::FRAC_1_SQRT_2 * Scalar::FRAC_1_SQRT_2;
-
-        let correct = vec![
-            ScaledState(ExtendedBasis::Superposition(vec![Zero, Zero]), half),
-            ScaledState(ExtendedBasis::Superposition(vec![One, Zero]), half),
-            ScaledState(ExtendedBasis::Superposition(vec![Zero, One]), -half),
-            ScaledState(ExtendedBasis::Superposition(vec![One, One]), -half),
-        ];
-        let attempt = basis_to_expand.clone().expand_qubits(expand_both);
-        assert!(correct.len() == attempt.len());
-        for state in correct {
-            assert!(attempt.contains(&state));
-        }
-    }
-
-    #[test]
-    fn test_hadamard_cnot_entanglement() {
-        let circuit = Circuit::new(2).h(0).cx(&[0], 1);
-        for _ in 0..1000 {
-            let mut sim = match SyntaxSimulator::build(circuit.clone()) {
-                Ok(sim) => sim,
-                Err(e) => panic!("Error building simulator: {}", e),
-            };
-            let result = sim.run();
-            assert!(result == 0b00 || result == 0b11);
-        }
     }
 }
